Fix for ICE in taskloop
Add missing case statement for end of Taskloop. Absence of this
case statement leads to some variables being marked as private
(instead of local). This prevents the allocation of these modules
on the stack and their further use in load or store statements.
This patch fixes https://github.com/flang-compiler/flang/issues/688.